requirements.txt
av==10.0.0
customtkinter==5.1.3
ffmpeg_python==0.2.0
numpy==1.23.5
opencv_python==4.7.0.72
Pillow==10.0.0
pytorch_lightning==2.0.2
pytorchvideo==0.1.5
scikit_learn==1.2.2
torch==2.0.0+cu118
torchmetrics==0.11.4
torchvision==0.15.1+cu118
ultralytics==8.0.147
